+++ src/mix_set_readable.ts
@@ -1,8 +1,9 @@
 import type { Readable, Updater } from 'svelte/store'
 import { assign } from '@ctx-core/object'
 import { get } from './get'
+import type { Writable_set_T, Writable_update_T } from './writable'
 export function mix_set_readable<Val extends unknown = unknown>(
-	store:Readable<Val>, set:Updater<Val>
+	store:Readable<Val>, set:Writable_set_T<Val>
 ):mix_set_readable_I<Val> {
 	return assign(store as mix_set_readable_I<Val>, {
 		set,
@@ -12,6 +13,6 @@ export function mix_set_readable<Val extends unknown = unknown>(
 	})
 }
 export interface mix_set_readable_I<Val extends unknown = unknown> extends Readable<Val> {
-	set:Updater<Val>
-	update(this: void, updater:Updater<Val>):void
+	set:Writable_set_T<Val>
+	update:Writable_update_T<Val>
 }