+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ctx-core/store
 
+## 24.4.1
+
+### Patch Changes
+
+- fix: mix_set_readable\$: set: Writable_set_T<Val>
+- - Writable_update_T<Val extends unknown = unknown>
+
 ## 24.4.0
 
 ### Minor Changes
@@ -18,7 +25,7 @@
 
 ### Minor Changes
 
-- - readable\\\$,derived\\\$,writable\\\$,Readable\\\$,Writable\$
+- - readable\\\\$,derived\\\\$,writable\\\\$,Readable\\\\$,Writable\$
 - \_readable_set_a1,\_readable_set_ctx: readable_fn argument
 
 ## 24.2.4